requirements.txt
Flask>=2.3.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.21.0
Werkzeug>=2.3.0
Jinja2>=3.1.0
MarkupSafe>=2.1.0
click>=8.1.0
itsdangerous>=2.1.0
blinker>=1.6.0
